Added the ability to configure which column to use when calculating HSK statistics.
Columns can either be specified by index (starting at 0), "all", or "sortField". The default behavior
is to use sortField, which matches the current implementation. It should be noted that sortField is not
actually the 'first' field as stated in the projects README. The config.md file is shown to the user
when they configure the plugin, the explanation may need some tweaking if it's not clear enough.

For some extra context, the Anki database schema can be found here: 
https://github.com/ankidroid/Anki-Android/wiki/Database-Structure