Configuration changes are now used when next bringing up the results page (instead of application restart).
The user facing Field numbers are now translated to the internal index, instead of relying on the user to
know they need to subtract one from what they see in the GUI.

Fields are now parsed correctly. The database returns a string delimited with US (ASCII 31) which is now parsed
correctly. Previously it was only grabbing the indexed character (ex. fields[1] would return the second character
in the string).